Burn down event-log formatter maintenance debt
Split the readable event-log formatter along the existing category boundaries so the CLI stays small while the combat/status formatting remains directly testable.

Constraint: Maintenance warning ledger requires fixed/follow-up accounting to match live scanner output.

Rejected: Force-adding ignored helper modules under scripts root | scripts/** is intentionally ignored except tracked support directories.

Confidence: high

Scope-risk: narrow

Directive: Keep event-log formatter helpers in tracked scripts/lib modules and lock CLI behavior with sample NDJSON fixtures before further splits.

Tested: npm.cmd test -- --watchAll=false --runTestsByPath scripts/__tests__/format-event-log.test.ts scripts/__tests__/maintenance-warning-ledger.test.ts --runInBand; npm.cmd run typecheck -- --pretty false; npm.cmd run format:check; npm.cmd run qc:validate; node scripts/qc/validate-maintenance-warning-ledger.mjs; npm.cmd run qc:app-completion -- --json; npm.cmd run verify:qc:maintenance; node scripts/maintenance/scan-maintenance.mjs --scope=scripts/format-event-log.py,scripts/lib/event_log_format_core.py,scripts/lib/event_log_format_combat.py,scripts/lib/event_log_format_status.py,scripts/lib/event_log_format_summaries.py --json; git diff --cached --check

Not-tested: Full historical event-log corpus parity beyond the sample NDJSON fixture.
