+++ docs/qc/mekstation-qc-validation-graph.json
@@ -1668,7 +1668,8 @@
         "docs/qc/maintenance-warning-ledger.json",
         "scripts/qc/validate-maintenance-warning-ledger.mjs",
         "2026-06-27 validate-bv architecture split moved the BattleMech BV calculation engine into focused calculator, crit-scan, ammo-resolution, weapon-resolution, known-unit, normalizer, and type modules; the original validate-bv.ts critical file-bloat finding is fixed, with three high file-bloat follow-ups tracked for the remaining validator submodules.",
-        "2026-06-28 maintenance release scope promoted to ready-with-scope after `verify:qc:maintenance` and `qc:app-completion -- --json` proved the src critical/high regression gate, repo-wide warning ledger accounting, and remaining follow-up/accepted warning debt are machine-checkable instead of hidden release blockers."
+        "2026-06-28 maintenance release scope promoted to ready-with-scope after `verify:qc:maintenance` and `qc:app-completion -- --json` proved the src critical/high regression gate, repo-wide warning ledger accounting, and remaining follow-up/accepted warning debt are machine-checkable instead of hidden release blockers.",
+        "2026-06-28 event-log formatter maintenance split moved prefix and summary helpers into tracked `scripts/lib/event_log_format_*.py` modules, added sample NDJSON fixture coverage in `scripts/__tests__/format-event-log.test.ts`, and fixed the `scripts/format-event-log.py` file-bloat warning without adding new actionable Python findings."
       ]
     },
     {
